requirements-api.txt
fastapi
uvicorn
python-multipart
torch
torchvision
timm
open-clip-torch
faiss-cpu
numpy
pandas
Pillow
requests
python-dotenv
